Create pythoncipher.py
Python Cipher is a digital Caesar cipher coded with Python and Pygame. It can run three different processes: encoding, decoding, and random decoding. The encoding process takes string input from the user as well as an integer value for the shift and returns the encoded message. The decoding process takes ciphertext from the user and uses a brute-force algorithm to decode the message. The random decoding process produces ciphertext and the user chooses shifts until the message is converted back to plaintext.